requirements.txt
fastapi>=0.95.0,<0.110.0
uvicorn>=0.21.1,<0.30.0
openai>=0.27.4,<1.0.0
python-multipart>=0.0.6,<0.1.0
python-dotenv>=1.0.0,<2.0.0
pydantic>=1.10.7,<2.0.0
google-cloud-storage>=2.10.0,<3.0.0
google-auth>=2.22.0,<3.0.0
streamlit==1.22.0
requests==2.28.2
python-dotenv==1.0.0
